Add files via upload
Code for a IOT Halloween crow with glowing eyes, based on an esp8266 board. 
Device opens an AP, which you can connect to and give credentials for actual wifi network. it then automatically connects to that wifi network and self-hosts a webpage under crow.local. You can change eye animations and color there.
Pins can be found below, documentation kind of missing at the moment.

For next helloween I'd like to implement a speaker on the crow, how exaclty or what will be played is not clear yet